Update Program.cs
Hello, my name is Marcus Vogado.
I'm a student and I bought your book, I noticed that in chapter 9 the example in the book is incomplete in the C# language. so I decided to put the studies into practice and updated the code. Thank you very much for the teachings.