Ch03 Countdown: Print numbers on the same line avoiding invalid integers
The current function `countdown()` prints integers equal and lesser than zero, which is not desired as shown in the book:

```text
> 3...2...1
```

I propose printing the numbers on the same line separated with ellipsis (...), except for number 1, which should be finished with a new line. Additionally, the base case should stop if the current number `i` is lesser than one.

-	modified:   01_countdown.py

```python
def countdown(i):
  # base case
  if i < 1:
    # Finish decreasing and printing the numbers
    print()
    return
  # recursive case
  else:
    # Separate each number with ellipsis (...), except for the number 1.
    element_separator = "..." if i > 1 else ""
    # Print the numbers on the same line
    print(i, end=element_separator)
    return countdown(i - 1)
```