Refactor: replace helper function with map
Since we've already been introduced to hash tables in previous chapter, I thought it would be a great way to use them in practice. Using map to store searched/visited items is effective, and it simplifies the code.

In the book's example, Python's set() was used for this matter. Using map is a Go way of doing that.